+++ gfl-plugin-bgc/gfl_plugin_bgc/__init__.py
@@ -1,0 +1,5 @@
+"""GFL BGC and Megasynthase Plugin package."""
+
+from gfl_plugin_bgc.plugin import BGCPlugin
+
+__all__ = ["BGCPlugin"]